functions/clf-ingesta-sap/requirements.txt
functions-framework==3.5.0
requests==2.31.0
google-auth==2.27.0
cloudevents==1.10.1